+++ packages/browser/src/index.ts
@@ -48,6 +48,7 @@ export { setActiveSpanInBrowser } from './tracing/setActiveSpan';
 export { spanStreamingIntegration } from './integrations/spanstreaming';
 export { fetchStreamPerformanceIntegration } from './integrations/fetchStreamPerformance';
 export { webVitalsIntegration } from './integrations/webVitals';
+export { bfcacheMetricsIntegration } from './integrations/bfcache';
 
 export type { RequestInstrumentationOptions } from './tracing/request';
 export {